unit testing for file IO development; reorganization of HDF5 IO function calls
