+++ unittests/testSuite.cpp
@@ -2,3 +2,4 @@
 #define BOOST_TEST_DYN_LINK
 #include <boost/test/unit_test.hpp>
 #include "potentialTests.h"
+#include "ioTests.h"
